# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EIAM API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/identity_source/delete/{id}": {
      "delete": {
        "tags": [
          "身份源管理"
        ],
        "summary": "删除身份源",
        "description": "权限标识: delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/identity_source/detail/{id}": {
      "get": {
        "tags": [
          "身份源管理"
        ],
        "summary": "身份源详情",
        "description": "权限标识: detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
    "/api/identity_source/save": {
      "post": {
        "tags": [
          "身份源管理"
        ],
        "summary": "保存身份源",
        "description": "权限标识: save",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveIdentitySourceReq"
              },
              "example": {
                "enabled": true,
                "id": 1,
                "ldap": "string",
                "local": "string",
                "name": "示例名称",
                "oidc": "string",
                "passkey": "string",
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/identity_source/test": {
      "post": {
        "tags": [
          "身份源管理"
        ],
        "summary": "测试身份源连接",
        "description": "权限标识: test",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveIdentitySourceReq"
              },
              "example": {
                "enabled": true,
                "id": 1,
                "ldap": "string",
                "local": "string",
                "name": "示例名称",
                "oidc": "string",
                "passkey": "string",
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/identity_source/toggle/{id}": {
      "post": {
        "tags": [
          "身份源管理"
        ],
        "summary": "切换启用状态",
        "description": "权限标识: toggle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "SaveIdentitySourceReq": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "ldap": {
            "type": "string",
            "description": "LDAP 专有配置",
            "example": "string"
          },
          "local": {
            "type": "string",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          },
          "oidc": {
            "type": "string",
            "example": "string"
          },
          "passkey": {
            "type": "string",
            "example": "string"
          },
          "type": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "example": {
          "enabled": true,
          "id": 1,
          "ldap": "string",
          "local": "string",
          "name": "示例名称",
          "oidc": "string",
          "passkey": "string",
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}